requirements.txt
Flask==2.2.2
Jinja2==3.1.2
discord.py==2.3.1